+++ R/analysis_celltypist.R
@@ -92,7 +92,9 @@
 #'
 #' @examples
 #' \dontrun{
-#' pbmc <- qs2::qs_read(file.path(Sys.getenv("SHENNONG_REAL_DATA_DIR"), "single-cell", "kotliarov_pbmc.qs2"))
+#' pbmc <- qs2::qs_read(file.path(
+#'   Sys.getenv("SHENNONG_REAL_DATA_DIR"), "single-cell", "kotliarov_pbmc.qs2"
+#' ))
 #' pbmc <- sn_run_cluster(pbmc, normalization_method = "seurat", verbose = FALSE)
 #' pbmc <- sn_run_celltypist(pbmc, model = "Immune_All_Low.pkl")
 #' head(colnames(pbmc[[]]))