+++ R/analysis_rogue.R
@@ -33,7 +33,9 @@
 #'
 #' @examples
 #' \dontrun{
-#' pbmc <- qs2::qs_read(file.path(Sys.getenv("SHENNONG_REAL_DATA_DIR"), "single-cell", "kotliarov_pbmc.qs2"))
+#' pbmc <- qs2::qs_read(file.path(
+#'   Sys.getenv("SHENNONG_REAL_DATA_DIR"), "single-cell", "kotliarov_pbmc.qs2"
+#' ))
 #' pbmc <- sn_run_cluster(pbmc, normalization_method = "seurat", verbose = FALSE)
 #' rogue_tbl <- sn_calculate_rogue(pbmc, cluster_by = "seurat_clusters")
 #' head(rogue_tbl)